update: new test architecture already working for courses
